JwtService class split
