+++ admin-service/src/main/java/ru/tbank/admin/exceptions/ApplicationRuntimeException.java
@@ -1,8 +1,5 @@
 package ru.tbank.admin.exceptions;
 
-import lombok.EqualsAndHashCode;
-
-@EqualsAndHashCode(callSuper = true)
 public class ApplicationRuntimeException extends RuntimeException {
 
     public ApplicationRuntimeException(String message) {
