+++ admin-service/src/main/java/ru/tbank/admin/exceptions/ChatNotFoundException.java
@@ -1,10 +1,8 @@
 package ru.tbank.admin.exceptions;
 
-import lombok.EqualsAndHashCode;
 import lombok.Getter;
 
 @Getter
-@EqualsAndHashCode(callSuper = true)
 public class ChatNotFoundException extends ApplicationRuntimeException {
 
     private final Long chatId;
