+++ admin-service/src/main/java/ru/tbank/admin/exceptions/InvalidFilterTypeException.java
@@ -1,8 +1,5 @@
 package ru.tbank.admin.exceptions;
 
-import lombok.EqualsAndHashCode;
-
-@EqualsAndHashCode(callSuper = true)
 public class InvalidFilterTypeException extends ApplicationRuntimeException {
 
     public InvalidFilterTypeException() {
